Initial commit of py-vmware tools
This commit adds a collection of python vmware tools in a single package to provide an easy entry point. It provides a set of capabilities that allow you to manage virtual machine creation, snapshots, deletion, migration and listing virtual machines. Additionally, it offers some capabilities to perform maintenance on ESXi hosts as well as resource rebalancing based on host utilization and some hard limitations.