Add capability to mount datastores
This commit adds a capability to mount datastores. Additionally, functions from migrate_host_vms are rolled into vmware_lib to provide a set of common capabilities for reuse across CLI interfaces. A capability to reboot a VM is added. Lastly, entry points are added for mounting datastores and rebooting VMs.