Add vm name to extraconfig and leverage vmware_lib for clone_vm
This commit updates clone_vm to use vmware_lib for connection
components. Additionally, extraconfig now gets populated with VM
name, which makes it possible to leverage boostrapping scripts that
expect to find this data for automatic naming of deployed VMs.
Additionally, a capability is added to explicitly identify the parent
folder of your folder target. Without this change it's possible to
select a different folder than expected when there is more than one
folder of that title in inventory. Additionally, without this change,
machines configured to find data about their name upon first boot, and
configure themselves with that name, will not work.